+++ src/clj/rems/auth/oidc.clj
@@ -86,7 +86,7 @@
         (find-first users/user-exists? (map second userid-attrs)))))
 
 (defn- upsert-user! [user]
-  (let [userid (:eppn user)]
+  (let [userid (:userid user)]
     (users/add-user-raw! userid user)
     user))
 
@@ -94,9 +94,9 @@
   ;; TODO all attributes could support :rename
   (let [userid (or (find-user id-data) (get-new-userid id-data))
         _ (assert userid (when (:log-authentication-details env) {:id-data id-data :user-info user-info}))
-        identity-base {:eppn userid
-                       :commonName (some (comp id-data keyword) (:oidc-name-attributes env))
-                       :mail (some (comp id-data keyword) (:oidc-email-attributes env))}
+        identity-base {:userid userid
+                       :name (some (comp id-data keyword) (:oidc-name-attributes env))
+                       :email (some (comp id-data keyword) (:oidc-email-attributes env))}
         extra-attributes (select-keys id-data (map (comp keyword :attribute) (:oidc-extra-attributes env)))
         user-info-attributes (select-keys user-info [:researcher-status-by])]
     (merge identity-base extra-attributes user-info-attributes)))
@@ -105,8 +105,8 @@
 (defn- validate-user! [user]
   ;; userid already checked
   (when-let [errors (seq (remove nil?
-                                 [(when (str/blank? (:commonName user)) :t.login.errors/name)
-                                  (when (str/blank? (:mail user)) :t.login.errors/email)]))]
+                                 [(when (str/blank? (:name user)) :t.login.errors/name)
+                                  (when (str/blank? (:email user)) :t.login.errors/email)]))]
     (throw (ex-info "Invalid user"
                     {:key :t.login.errors/invalid-user
                      :args errors
@@ -116,7 +116,7 @@
   (let [user (get-user-attributes id-data user-info)
         _ (validate-user! user)
         user (upsert-user! user)]
-    (save-user-mappings! id-data (:eppn user))
+    (save-user-mappings! id-data (:userid user))
     user))
 
 (defn oidc-callback [request]
@@ -174,25 +174,37 @@
 
 (defn- oidc-revoke [token]
   (when token
-    (when-let [endpoint (:revocation_endpoint oidc-configuration)]
-      (let [response (http/post endpoint
-                                {:basic-auth [(getx env :oidc-client-id)
-                                              (getx env :oidc-client-secret)]
-                                 :form-params {:token token}
-                                 :throw-exceptions false})]
-        (when-not (= 200 (:status response))
-          (log/error "received HTTP status" (:status response) "from" endpoint))))))
+    (let [endpoint (:revocation_endpoint oidc-configuration)]
+      (when (:log-authentication-details env)
+        (log/info "revoking token" endpoint))
 
-; TODO Logout. Federated or not?
-; TODO Silent login when we have a new session, but user has logged in to auth provider
+      (when endpoint
+        (let [response (http/post endpoint
+                                  {:basic-auth [(getx env :oidc-client-id)
+                                                (getx env :oidc-client-secret)]
+                                   :form-params {:token token}
+                                   :throw-exceptions false})]
+          (when (:log-authentication-details env)
+            (log/info "performed revocation" (:status response)))
+
+          (when-not (= 200 (:status response))
+            (log/error "received HTTP status" (:status response) "from" endpoint)))))))
 
 (defroutes routes
   (GET "/oidc-login" _req (redirect (login-url)))
   (GET "/oidc-logout" req
-    (let [session (get req :session)]
+    (let [session (get req :session)
+          redirect-url (:oidc-logout-redirect-url env)]
       (when (:log-authentication-details env)
         (log/info "logging out" (:identity session)))
-      (oidc-revoke (:access-token session))
-      (assoc (redirect "/") :session (dissoc session :identity :access-token))))
+
+      (when (:oidc-perform-revoke-in-logout env)
+        (oidc-revoke (:access-token session)))
+
+      (when (:log-authentication-details env)
+        (log/info "redirecting to" redirect-url))
+
+      (assoc (redirect redirect-url)
+             :session (dissoc session :identity :access-token))))
   (GET "/oidc-callback" req (oidc-callback req)))