+++ src/clj/rems/db/test_data_users.clj
@@ -12,10 +12,10 @@
    :expirer-bot expirer-bot/bot-userid})
 
 (def +bot-user-data+
-  {approver-bot/bot-userid {:eppn approver-bot/bot-userid :commonName "Approver Bot"}
-   bona-fide-bot/bot-userid {:eppn bona-fide-bot/bot-userid :commonName "Bona Fide Bot"}
-   rejecter-bot/bot-userid {:eppn rejecter-bot/bot-userid :commonName "Rejecter Bot"}
-   expirer-bot/bot-userid {:eppn expirer-bot/bot-userid :commonName "Expirer Bot"}})
+  {approver-bot/bot-userid {:userid approver-bot/bot-userid :name "Approver Bot"}
+   bona-fide-bot/bot-userid {:userid bona-fide-bot/bot-userid :name "Bona Fide Bot"}
+   rejecter-bot/bot-userid {:userid rejecter-bot/bot-userid :name "Rejecter Bot"}
+   expirer-bot/bot-userid {:userid expirer-bot/bot-userid :name "Expirer Bot"}})
 
 (def +fake-users+
   {:applicant1 "alice"
@@ -33,19 +33,19 @@
    :jillsmith "jillsmith"})
 
 (def +fake-user-data+
-  {"developer" {:eppn "developer" :mail "developer@example.com" :commonName "Developer" :nickname "The Dev"}
-   "alice" {:eppn "alice" :mail "alice@example.com" :commonName "Alice Applicant" :organizations [{:organization/id "default"}] :nickname "In Wonderland" :researcher-status-by "so"}
-   "malice" {:eppn "malice" :mail "malice@example.com" :commonName "Malice Applicant" :twinOf "alice" :other "Attribute Value" :mappings {"alt-id" "malice-alt-id"}}
-   "handler" {:eppn "handler" :mail "handler@example.com" :commonName "Hannah Handler" :mappings {"alt-id" "handler-alt-id"}}
-   "carl" {:eppn "carl" :mail "carl@example.com" :commonName "Carl Reviewer" :mappings {"alt-id" "carl-alt-id"}}
-   "elsa" {:eppn "elsa" :mail "elsa@example.com" :commonName "Elsa Roleless" :mappings {"alt-id" "elsa-alt-id"}}
-   "frank" {:eppn "frank" :mail "frank@example.com" :commonName "Frank Roleless" :organizations [{:organization/id "frank"}]}
-   "organization-owner1" {:eppn "organization-owner1" :mail "organization-owner1@example.com" :commonName "Organization Owner 1"}
-   "organization-owner2" {:eppn "organization-owner2" :mail "organization-owner2@example.com" :commonName "Organization Owner 2" :organizations [{:organization/id "organization2"}]}
-   "owner" {:eppn "owner" :mail "owner@example.com" :commonName "Owner"}
-   "reporter" {:eppn "reporter" :mail "reporter@example.com" :commonName "Reporter"}
-   "johnsmith" {:eppn "johnsmith" :commonName "John Smith" :mail "john.smith@example.com" :mappings {"identity1" "johnsmith" "identity2" "smith"}}
-   "jillsmith" {:eppn "jillsmith" :commonName "Jill Smith" :mail "jill.smith@example.com" :mappings {"identity1" "jillsmith" "identity2" "smith"}}})
+  {"developer" {:userid "developer" :email "developer@example.com" :name "Developer" :nickname "The Dev"}
+   "alice" {:userid "alice" :email "alice@example.com" :name "Alice Applicant" :organizations [{:organization/id "default"}] :nickname "In Wonderland" :researcher-status-by "so"}
+   "malice" {:userid "malice" :email "malice@example.com" :name "Malice Applicant" :twinOf "alice" :other "Attribute Value" :mappings {"alt-id" "malice-alt-id"}}
+   "handler" {:userid "handler" :email "handler@example.com" :name "Hannah Handler" :mappings {"alt-id" "handler-alt-id"}}
+   "carl" {:userid "carl" :email "carl@example.com" :name "Carl Reviewer" :mappings {"alt-id" "carl-alt-id"}}
+   "elsa" {:userid "elsa" :email "elsa@example.com" :name "Elsa Roleless" :mappings {"alt-id" "elsa-alt-id"}}
+   "frank" {:userid "frank" :email "frank@example.com" :name "Frank Roleless" :organizations [{:organization/id "frank"}]}
+   "organization-owner1" {:userid "organization-owner1" :email "organization-owner1@example.com" :name "Organization Owner 1"}
+   "organization-owner2" {:userid "organization-owner2" :email "organization-owner2@example.com" :name "Organization Owner 2" :organizations [{:organization/id "organization2"}]}
+   "owner" {:userid "owner" :email "owner@example.com" :name "Owner"}
+   "reporter" {:userid "reporter" :email "reporter@example.com" :name "Reporter"}
+   "johnsmith" {:userid "johnsmith" :name "John Smith" :email "john.smith@example.com" :mappings {"identity1" "johnsmith" "identity2" "smith"}}
+   "jillsmith" {:userid "jillsmith" :name "Jill Smith" :email "jill.smith@example.com" :mappings {"identity1" "jillsmith" "identity2" "smith"}}})
 
 (def +fake-id-data+
   {"developer" {:sub "developer" :email "developer@example.com" :name "Developer" :nickname "The Dev"}
@@ -126,15 +126,15 @@
    :reporter "RDdomainreporter@funet.fi"})
 
 (def +demo-user-data+
-  {"RDapplicant1@funet.fi" {:eppn "RDapplicant1@funet.fi" :mail "RDapplicant1.test@test_example.org" :commonName "RDapplicant1 REMSDEMO1" :organizations [{:organization/id "default"}]}
-   "RDapplicant2@funet.fi" {:eppn "RDapplicant2@funet.fi" :mail "RDapplicant2.test@test_example.org" :commonName "RDapplicant2 REMSDEMO"}
-   "RDapprover1@funet.fi" {:eppn "RDapprover1@funet.fi" :mail "RDapprover1.test@rems_example.org" :commonName "RDapprover1 REMSDEMO"}
-   "RDapprover2@funet.fi" {:eppn "RDapprover2@funet.fi" :mail "RDapprover2.test@rems_example.org" :commonName "RDapprover2 REMSDEMO"}
-   "RDreview@funet.fi" {:eppn "RDreview@funet.fi" :mail "RDreview.test@rems_example.org" :commonName "RDreview REMSDEMO"}
-   "RDowner@funet.fi" {:eppn "RDowner@funet.fi" :mail "RDowner.test@test_example.org" :commonName "RDowner REMSDEMO"}
-   "RDorganizationowner1@funet.fi" {:eppn "RDorganizationowner1@funet.fi" :mail "RDorganizationowner1.test@test_example.org" :commonName "RDorganizationowner1 REMSDEMO" :organizations [{:organization/id "organization1"}]}
-   "RDorganizationowner2@funet.fi" {:eppn "RDorganizationowner2@funet.fi" :mail "RDorganizationowner2.test@test_example.org" :commonName "RDorganizationowner2 REMSDEMO" :organizations [{:organization/id "organization2"}]}
-   "RDdomainreporter@funet.fi" {:eppn "RDdomainreporter@funet.fi" :mail "RDdomainreporter.test@test_example.org" :commonName "RDdomainreporter REMSDEMO"}})
+  {"RDapplicant1@funet.fi" {:userid "RDapplicant1@funet.fi" :email "RDapplicant1.test@test_example.org" :name "RDapplicant1 REMSDEMO1" :organizations [{:organization/id "default"}]}
+   "RDapplicant2@funet.fi" {:userid "RDapplicant2@funet.fi" :email "RDapplicant2.test@test_example.org" :name "RDapplicant2 REMSDEMO"}
+   "RDapprover1@funet.fi" {:userid "RDapprover1@funet.fi" :email "RDapprover1.test@rems_example.org" :name "RDapprover1 REMSDEMO"}
+   "RDapprover2@funet.fi" {:userid "RDapprover2@funet.fi" :email "RDapprover2.test@rems_example.org" :name "RDapprover2 REMSDEMO"}
+   "RDreview@funet.fi" {:userid "RDreview@funet.fi" :email "RDreview.test@rems_example.org" :name "RDreview REMSDEMO"}
+   "RDowner@funet.fi" {:userid "RDowner@funet.fi" :email "RDowner.test@test_example.org" :name "RDowner REMSDEMO"}
+   "RDorganizationowner1@funet.fi" {:userid "RDorganizationowner1@funet.fi" :email "RDorganizationowner1.test@test_example.org" :name "RDorganizationowner1 REMSDEMO" :organizations [{:organization/id "organization1"}]}
+   "RDorganizationowner2@funet.fi" {:userid "RDorganizationowner2@funet.fi" :email "RDorganizationowner2.test@test_example.org" :name "RDorganizationowner2 REMSDEMO" :organizations [{:organization/id "organization2"}]}
+   "RDdomainreporter@funet.fi" {:userid "RDdomainreporter@funet.fi" :email "RDdomainreporter.test@test_example.org" :name "RDdomainreporter REMSDEMO"}})
 
 (def +demo-id-data+
   {"RDapplicant1@funet.fi" {:sub "RDapplicant1@funet.fi" :email "RDapplicant1.test@test_example.org" :name "RDapplicant1 REMSDEMO1" :organizations [{:organization/id "default"}]}
@@ -162,12 +162,12 @@
    :owner "BACZQAPVWBDJ2OXLKT2WWW5LT5LV6YR4"})
 
 (def +oidc-user-data+
-  {"WHFS36UEZD6TNURJ76WYLSVDCUUENOOF" {:eppn "WHFS36UEZD6TNURJ76WYLSVDCUUENOOF" :mail "RDapplicant1@mailinator.com" :commonName "RDapplicant1 REMSDEMO1" :organizations [{:organization/id "default"}]}
-   "C567LI5QAACWKC7YYA74BJ2X7DH7EEYI" {:eppn "C567LI5QAACWKC7YYA74BJ2X7DH7EEYI" :mail "RDapplicant2@mailinator.com" :commonName "RDapplicant2 REMSDEMO"}
-   "EKGFNAAGCHIQ5ERUUFS2RCZ44IHYZPEA" {:eppn "EKGFNAAGCHIQ5ERUUFS2RCZ44IHYZPEA" :mail "RDapprover1@mailinator.com" :commonName "RDapprover1 REMSDEMO"}
-   "7R3JYB32PL3EPVD34RWIAWDZSEOXW4OQ" {:eppn "7R3JYB32PL3EPVD34RWIAWDZSEOXW4OQ" :mail "RDapprover2@mailinator.com" :commonName "RDapprover2 REMSDEMO"}
-   "F3OJL757ACT4QXVXZZ4F7VG6HQGBEC4M" {:eppn "F3OJL757ACT4QXVXZZ4F7VG6HQGBEC4M" :mail "RDreview@mailinator.com" :commonName "RDreview REMSDEMO"}
-   "JOBDHBMX4EFXQC5IPQVXPP4FFWJ6XQYL" {:eppn "JOBDHBMX4EFXQC5IPQVXPP4FFWJ6XQYL" :mail "RDdomainreporter@mailinator.com" :commonName "RDdomainreporter REMSDEMO"}
-   "W6OKPQGANG6QK54GRF7AOOGMZL7M6IVH" {:eppn "W6OKPQGANG6QK54GRF7AOOGMZL7M6IVH" :mail "RDorganizationowner1@mailinator.com" :commonName "RDorganizationowner1 REMSDEMO" :organizations [{:organization/id "organization1"}]}
-   "D4ZJM7XNXKGFQABRQILDI6EYHLJRLYSF" {:eppn "D4ZJM7XNXKGFQABRQILDI6EYHLJRLYSF" :mail "RDorganizationowner2@mailinator.com" :commonName "RDorganizationowner2 REMSDEMO" :organizations [{:organization/id "organization2"}]}
-   "BACZQAPVWBDJ2OXLKT2WWW5LT5LV6YR4" {:eppn "BACZQAPVWBDJ2OXLKT2WWW5LT5LV6YR4" :mail "RDowner@mailinator.com" :commonName "RDowner REMSDEMO"}})
+  {"WHFS36UEZD6TNURJ76WYLSVDCUUENOOF" {:userid "WHFS36UEZD6TNURJ76WYLSVDCUUENOOF" :email "RDapplicant1@mailinator.com" :name "RDapplicant1 REMSDEMO1" :organizations [{:organization/id "default"}]}
+   "C567LI5QAACWKC7YYA74BJ2X7DH7EEYI" {:userid "C567LI5QAACWKC7YYA74BJ2X7DH7EEYI" :email "RDapplicant2@mailinator.com" :name "RDapplicant2 REMSDEMO"}
+   "EKGFNAAGCHIQ5ERUUFS2RCZ44IHYZPEA" {:userid "EKGFNAAGCHIQ5ERUUFS2RCZ44IHYZPEA" :email "RDapprover1@mailinator.com" :name "RDapprover1 REMSDEMO"}
+   "7R3JYB32PL3EPVD34RWIAWDZSEOXW4OQ" {:userid "7R3JYB32PL3EPVD34RWIAWDZSEOXW4OQ" :email "RDapprover2@mailinator.com" :name "RDapprover2 REMSDEMO"}
+   "F3OJL757ACT4QXVXZZ4F7VG6HQGBEC4M" {:userid "F3OJL757ACT4QXVXZZ4F7VG6HQGBEC4M" :email "RDreview@mailinator.com" :name "RDreview REMSDEMO"}
+   "JOBDHBMX4EFXQC5IPQVXPP4FFWJ6XQYL" {:userid "JOBDHBMX4EFXQC5IPQVXPP4FFWJ6XQYL" :email "RDdomainreporter@mailinator.com" :name "RDdomainreporter REMSDEMO"}
+   "W6OKPQGANG6QK54GRF7AOOGMZL7M6IVH" {:userid "W6OKPQGANG6QK54GRF7AOOGMZL7M6IVH" :email "RDorganizationowner1@mailinator.com" :name "RDorganizationowner1 REMSDEMO" :organizations [{:organization/id "organization1"}]}
+   "D4ZJM7XNXKGFQABRQILDI6EYHLJRLYSF" {:userid "D4ZJM7XNXKGFQABRQILDI6EYHLJRLYSF" :email "RDorganizationowner2@mailinator.com" :name "RDorganizationowner2 REMSDEMO" :organizations [{:organization/id "organization2"}]}
+   "BACZQAPVWBDJ2OXLKT2WWW5LT5LV6YR4" {:userid "BACZQAPVWBDJ2OXLKT2WWW5LT5LV6YR4" :email "RDowner@mailinator.com" :name "RDowner REMSDEMO"}})